refactor(nms): move fbc lib logging
