+++ nms/fbc_js_core/sequelize_models/sequelizeConfig.js
@@ -24,7 +24,7 @@ const MYSQL_PASS = process.env.MYSQL_PASS || '';
 const MYSQL_DB = process.env.MYSQL_DB || 'cxl';
 const MYSQL_DIALECT = process.env.MYSQL_DIALECT || 'mysql';
 
-const logger = require('../../fbc_js_core/logging').getLogger(module);
+const logger = require('../../shared/logging').getLogger(module);
 
 let ssl_required = false;
 let CAcert = process.env.CA_FILE;
