+++ nms/server/apicontroller/auditLoggingDecorator.js
@@ -21,7 +21,7 @@ const url = require('url');
 import pathToRegexp from 'path-to-regexp';
 
 import {AuditLogEntry} from '../../fbc_js_core/sequelize_models';
-const logger = require('../../fbc_js_core/logging').getLogger(module);
+const logger = require('../../shared/logging').getLogger(module);
 
 const defaultResolver = (req: FBCNMSRequest, type: string) => {
   const {search} = url.parse(req.originalUrl);
