+++ nms/server/auth/express.js
@@ -17,7 +17,7 @@
 import bcrypt from 'bcryptjs';
 import express from 'express';
 import expressOnboarding from './expressOnboarding';
-import logging from '../../fbc_js_core/logging';
+import logging from '../../shared/logging';
 import passport from 'passport';
 import staticDist from '../../fbc_js_core/webpack_config/staticDist';
 import {AccessRoles} from '../../shared/roles';
