+++ nms/server/auth/strategies/OrganizationSamlStrategy.js
@@ -23,7 +23,7 @@ import {injectOrganizationParams} from '../organization';
 
 import crypto from 'crypto';
 
-const logger = require('../../../fbc_js_core/logging').getLogger(module);
+const logger = require('../../../shared/logging').getLogger(module);
 
 type Config = {
   urlPrefix: string,
