+++ nms/server/network/routes.js
@@ -31,7 +31,7 @@ import {CWF, FEG, FEG_LTE, LTE, XWFM} from '../../fbc_js_core/types/network';
 import {access} from '../auth/access';
 import {difference} from 'lodash';
 
-const logger = require('../../fbc_js_core/logging').getLogger(module);
+const logger = require('../../shared/logging').getLogger(module);
 
 const router: express.Router<FBCNMSRequest, ExpressResponse> = express.Router();
