+++ nms/shared/logging.js
